requirements.txt
asgiref==3.8.1
asttokens==2.4.1
autopep8==2.0.0
certifi==2024.8.30
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==3.1.0
commonmark==0.9.1
dask==2023.3.1
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.1.3
django-cors-headers==4.6.0
django-extensions==3.2.3
django-filter==24.3
django-import-export==3.1.0
django-seed==0.3.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djhtml==1.5.2
et_xmlfile==2.0.0
executing==2.1.0
Faker==33.1.0
faker-commerce==1.0.3
fsspec==2024.10.0
gunicorn==20.1.0
idna==3.10
ipdb==0.13.13
ipython==8.30.0
isort==5.10.1
jedi==0.19.2
Jinja2==3.1.4
locket==1.0.0
Markdown==3.7
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
numpy==2.1.3
odfpy==1.4.1
openpyxl==3.0.10
packaging==24.2
pandas==1.5.3
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
Pillow==9.3.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
PyExcelerate==0.10.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-decouple==3.6
pytz==2024.2
PyYAML==6.0.2
requests==2.28.1
rich==12.6.0
ruff==0.0.254
six==1.16.0
sqlparse==0.5.2
stack-data==0.6.3
tablib==3.7.0
tomli==2.2.1
toolz==1.0.0
toposort==1.10
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==1.26.20
wcwidth==0.2.13
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
